Refactor fee estimation testing
