--- TODO
@@ -1,2 +1,0 @@
-decide where to test fee estimate response
-and if to test it at all
